OneDrive/Escritorio/apps/por hacer/Scrasping dropi/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic-settings>=2.4.0
sqlalchemy>=2.0.0
PyJWT>=2.8.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
playwright>=1.46.0
httpx>=0.27.0
apscheduler>=3.10.4
python-multipart>=0.0.9
pytest>=8.3.0
pytest-asyncio>=0.23.0
